+++ lib/dio_curl_interceptor.dart
@@ -3,9 +3,11 @@ library dio_curl_interceptor;
 export 'package:colored_logger/colored_logger.dart';
 
 export 'src/data/curl_response_cache.dart';
-export 'src/core/curl_utils.dart';
+export 'src/core/utils/curl_utils.dart';
 export 'src/core/types.dart';
+export 'src/core/inspector.dart';
 export 'src/interceptors/dio_curl_interceptor_base.dart';
 export 'src/options/cache_options.dart';
 export 'src/options/curl_options.dart';
+export 'src/options/inspector_options.dart';
 export 'src/ui/curl_viewer_popup.dart';