+++ lib/src/core/utils/inspector_utils.dart
@@ -1,0 +1,9 @@
+import '../../options/inspector_options.dart';
+
+class InspectorUtils {
+  InspectorUtils({
+    this.inspectorOptions,
+  });
+
+  final InspectorOptions? inspectorOptions;
+}