+++ test/curl_helpers_test.dart
@@ -1,6 +1,6 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:dio_curl_interceptor/src/core/curl_helpers.dart';
+import 'package:dio_curl_interceptor/src/core/helpers.dart';
 import 'package:dio_curl_interceptor/src/options/curl_options.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group('Pretty', () {